+++ modules/mcp_tools_layout_builder/src/Service/LayoutBuilderService.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Drupal\mcp_tools_layout_builder\Service;
 
+use Drupal\Component\Uuid\UuidInterface;
 use Drupal\Core\Block\BlockManagerInterface;
 use Drupal\Core\Entity\EntityDisplayRepositoryInterface;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
@@ -23,6 +24,7 @@ class LayoutBuilderService {
     protected EntityDisplayRepositoryInterface $entityDisplayRepository,
     protected LayoutPluginManagerInterface $layoutPluginManager,
     protected BlockManagerInterface $blockManager,
+    protected UuidInterface $uuid,
     protected AccessManager $accessManager,
     protected AuditLogger $auditLogger,
   ) {}
@@ -489,7 +491,7 @@ class LayoutBuilderService {
     }
 
     try {
-      $componentUuid = \Drupal::service('uuid')->generate();
+      $componentUuid = $this->uuid->generate();
       $component = new SectionComponent($componentUuid, $region, ['id' => $blockId]);
       $section->appendComponent($component);